added metaplugin_proxy class, aka zzub_plugin_t in the c and plugin apis and updated c methods and events to use proxy objects instead of player&ints for addressing plugins

+++ src/libzzub/connections.cpp
@@ -231,7 +231,7 @@ int midi_connection::get_midi_device(zzub::song& player, int plugin_id, std::str
 	std::vector<string>::iterator i = find(midiouts.names.begin(), midiouts.names.end(), name);
 	if (i == midiouts.names.end()) return -1;
 	// we really need to find the _global_ device index, not the target machine one.... ? ?? ???
-	return i - midiouts.names.begin();
+	return (int)(i - midiouts.names.begin());
 }
 
 } // namespace zzub
